+++ http/service.go
@@ -9,6 +9,7 @@ import (
 	"errors"
 	"expvar"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"log"
 	"net"
@@ -56,8 +57,8 @@ type Store interface {
 	// Stats returns stats on the Store.
 	Stats() (map[string]interface{}, error)
 
-	// Backup returns a byte slice representing a backup of the node state.
-	Backup(leader bool, f store.BackupFormat) ([]byte, error)
+	// Backup writes backup of the node state to dst
+	Backup(leader bool, f store.BackupFormat, dst io.Writer) error
 }
 
 // CredentialStore is the interface credential stores must support.
@@ -391,17 +392,12 @@ func (s *Service) handleBackup(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	b, err := s.store.Backup(!noLeader, bf)
+	err = s.store.Backup(!noLeader, bf, w)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
 
-	_, err = w.Write(b)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
 	s.lastBackup = time.Now()
 }